multi-factor-system/requirements.txt
# 多因子选股系统依赖
# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# 数据获取
akshare>=1.12.0

# 机器学习/统计
scikit-learn>=1.3.0
statsmodels>=0.14.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 配置与日志
pyyaml>=6.0
tqdm>=4.66.0
loguru>=0.7.0

# Jupyter支持
jupyter>=1.0
ipykernel>=6.25

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0
